adds aged items type
also refactored some of the 'magical' numbers into constants
in an attempt to provide better context - can probably still be improved.

the next step of refactoring should be to wrap the normal Item class
in something that implentes #age with the aging logic of normal items.
This would allow us to replace everything in #update_quality